# uozhou/packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Restaurant Operations API",
    "version": "1.0.0",
    "description": "Odyssey 餐厅运营系统 API - JWT 认证"
  },
  "components": {
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "invalid_request"
          },
          "message": {
            "type": "string",
            "example": "请求参数错误"
          }
        },
        "required": [
          "error"
        ]
      },
...
      "UserRole": {
        "type": "string",
        "enum": [
          "admin",
          "staff",
          "customer"
        ]
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "phone": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "customerId": {
            "type": "integer",
            "nullable": true
          }
        },
        "required": [
          "id",
          "phone",
          "name",
          "role"
        ]
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          },
          "expiresIn": {
            "type": "integer"
          }
        },
        "required": [
          "user",
          "accessToken",
          "refreshToken",
          "expiresIn"
        ]
      },
      "LoginRequest": {
        "type": "object",
        "properties": {
          "phone": {
            "type": "string",
            "pattern": "^1[3-9]\\d{9}$"
          },
          "code": {
            "type": "string",
            "minLength": 6,
            "maxLength": 6
          }
        },
        "required": [
          "phone",
          "code"
        ]
      },
...
      "SuccessMessage": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "操作成功"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
      "MenuItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "price": {
            "type": "integer",
            "description": "以分为单位"
          },
          "image": {
            "type": "string",
            "nullable": true
          },
          "categoryId": {
            "type": "integer"
          },
          "categoryName": {
            "type": "string",
            "nullable": true
          },
          "available": {
            "type": "boolean"
          },
          "stock": {
            "type": "integer",
            "description": "-1 表示无限库存"
          },
          "sortOrder": {
            "type": "integer",
            "nullable": true
          }
        },
        "required": [
          "id",
          "name",
          "price",
          "categoryId",
          "available",
          "stock"
        ]
      },
      "CreateMenuItemRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "image": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "available": {
            "type": "boolean"
          },
          "stock": {
            "type": "integer",
            "minimum": -1
          },
          "sortOrder": {
            "type": "integer"
          }
        },
        "required": [
          "name",
          "price",
          "categoryId"
        ]
      },
      "UpdateMenuItemRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "description": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "image": {
            "type": "string"
          },
          "categoryId": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "available": {
            "type": "boolean"
          },
          "stock": {
            "type": "integer",
            "minimum": -1
          },
          "sortOrder": {
            "type": "integer"
          }
        }
      },
      "OrderStatus": {
        "type": "string",
        "enum": [
          "pending",
          "confirmed",
          "ready",
          "completed",
          "cancelled"
        ]
      },
      "Order": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "orderNo": {
            "type": "string"
          },
          "customerId": {
            "type": "integer"
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "total": {
            "type": "integer"
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "confirmedAt": {
            "type": "string",
            "nullable": true
          },
          "readyAt": {
            "type": "string",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "nullable": true
          },
          "cancelledAt": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "id",
          "orderNo",
          "customerId",
          "status",
          "total",
          "createdAt"
        ]
      },
      "OrderListResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Order"
            }
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "pageSize": {
            "type": "integer"
          }
        },
        "required": [
          "data",
          "total",
          "page",
          "pageSize"
        ]
      },
      "CreateOrderRequest": {
        "type": "object",
        "properties": {
          "customerId": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "menuItemId": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": true
                },
                "quantity": {
                  "type": "integer",
                  "minimum": 0,
                  "exclusiveMinimum": true
                }
              },
              "required": [
                "menuItemId",
                "quantity"
              ]
            },
            "minItems": 1
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "customerId",
          "items"
        ]
      },
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "orderId": {
            "type": "integer"
          },
          "menuItemId": {
            "type": "integer"
          },
          "menuItemName": {
            "type": "string"
          },
          "quantity": {
            "type": "integer"
          },
          "unitPrice": {
            "type": "integer"
          },
          "subtotal": {
            "type": "integer"
          }
        },
        "required": [
          "id",
          "orderId",
          "menuItemId",
          "menuItemName",
          "quantity",
          "unitPrice",
          "subtotal"
        ]
      },
      "Customer": {
        "type": "object",
        "nullable": true,
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "avatar": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "phone",
          "createdAt",
          "updatedAt"
        ]
      },
      "OrderWithItems": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Order"
          },
          {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/OrderItem"
                }
              },
              "customer": {
                "$ref": "#/components/schemas/Customer"
              }
            }
          }
        ]
      },
      "UpdateOrderRequest": {
        "type": "object",
        "properties": {
          "notes": {
            "type": "string"
          }
        }
      },
      "CreateCustomerRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1
          },
          "phone": {
            "type": "string",
            "pattern": "^1[3-9]\\d{9}$"
          },
          "avatar": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "phone"
        ]
      },
...
      "BusinessSettings": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "openingTime": {
            "type": "string"
          },
          "closingTime": {
            "type": "string"
          },
          "autoAcceptOrders": {
            "type": "boolean"
          },
          "estimatedPrepTime": {
            "type": "integer"
          },
          "serviceAvailable": {
            "type": "boolean"
          },
          "updatedAt": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "openingTime",
          "closingTime",
          "autoAcceptOrders",
          "estimatedPrepTime",
          "serviceAvailable",
          "updatedAt"
        ]
      },
      "UpdateBusinessSettingsRequest": {
        "type": "object",
        "properties": {
          "openingTime": {
            "type": "string",
            "pattern": "^\\d{2}:\\d{2}$"
          },
          "closingTime": {
            "type": "string",
            "pattern": "^\\d{2}:\\d{2}$"
          },
          "autoAcceptOrders": {
            "type": "boolean"
          },
          "estimatedPrepTime": {
            "type": "integer",
            "minimum": 1,
            "maximum": 120
          },
          "serviceAvailable": {
            "type": "boolean"
          }
        }
      }
    },
...
  },
  "paths": {
...
    "/api/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "手机号验证码登录",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "登录成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "400": {
            "description": "验证码错误或已过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/menu-items": {
      "get": {
        "tags": [
          "Menu"
        ],
        "summary": "获取菜品列表",
        "responses": {
          "200": {
            "description": "菜品列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MenuItem"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Menu"
        ],
        "summary": "新增菜品（admin）",
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMenuItemRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItem"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/menu-items/{id}": {
      "patch": {
        "tags": [
          "Menu"
        ],
        "summary": "修改菜品（admin）",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "example": "1"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateMenuItemRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "更新成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MenuItem"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "菜品不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Menu"
        ],
        "summary": "删除菜品（admin）",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "example": "1"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessMessage"
                }
              }
            }
          },
          "404": {
            "description": "菜品不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "订单列表（admin/staff）",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "$ref": "#/components/schemas/OrderStatus"
            },
            "required": false,
            "name": "status",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "1"
            },
            "required": false,
            "name": "page",
            "in": "query"
          },
          {
            "schema": {
              "type": "string",
              "example": "20"
            },
            "required": false,
            "name": "pageSize",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "startDate",
            "in": "query"
          },
          {
            "schema": {
              "type": "string"
            },
            "required": false,
            "name": "endDate",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "订单列表",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderListResponse"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "创建订单",
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/orders/{id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "获取订单详情",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "example": "1"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "responses": {
          "200": {
            "description": "订单详情",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderWithItems"
                }
              }
            }
          },
          "404": {
            "description": "订单不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Orders"
        ],
        "summary": "修改订单备注（禁止改 status）",
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "pattern": "^\\d+$",
              "example": "1"
            },
            "required": true,
            "name": "id",
            "in": "path"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "已更新",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "禁止改 status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/customers": {
      "get": {
        "tags": [
          "Customers"
        ],
        "summary": "客户列表（admin/staff）",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "客户列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Customer"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Customers"
        ],
        "summary": "新建客户",
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCustomerRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Customer"
                }
              }
            }
          },
          "400": {
            "description": "参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "获取业务设置",
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "业务设置",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessSettings"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "更新业务设置",
        "security": [
          {
            "Bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBusinessSettingsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "已更新",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BusinessSettings"
                }
              }
            }
          }
        }
      }
    }
  }
}