# uozhou/packages/api-client/src/generated/model/businessSettings.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Restaurant Operations API
 * Odyssey 餐厅运营系统 API - JWT 认证
 * OpenAPI spec version: 1.0.0
 */

export interface BusinessSettings {
  id: number;
  openingTime: string;
  closingTime: string;
  autoAcceptOrders: boolean;
  estimatedPrepTime: number;
  serviceAvailable: boolean;
  updatedAt: string;
}