# uozhou/packages/api-client/src/generated/model/menuItem.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Restaurant Operations API
 * Odyssey 餐厅运营系统 API - JWT 认证
 * OpenAPI spec version: 1.0.0
 */

export interface MenuItem {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** 以分为单位 */
  price: number;
  /** @nullable */
  image?: string | null;
  categoryId: number;
  /** @nullable */
  categoryName?: string | null;
  available: boolean;
  /** -1 表示无限库存 */
  stock: number;
  /** @nullable */
  sortOrder?: number | null;
}