# uozhou/packages/api-client/src/generated/model/order.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Restaurant Operations API
 * Odyssey 餐厅运营系统 API - JWT 认证
 * OpenAPI spec version: 1.0.0
 */
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  orderNo: string;
  customerId: number;
  /** @nullable */
  customerName?: string | null;
  status: OrderStatus;
  total: number;
  /** @nullable */
  notes?: string | null;
  createdAt: string;
  /** @nullable */
  confirmedAt?: string | null;
  /** @nullable */
  readyAt?: string | null;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  cancelledAt?: string | null;
}