# uozhou/packages/api-client/src/generated/model/orderListResponse.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Restaurant Operations API
 * Odyssey 餐厅运营系统 API - JWT 认证
 * OpenAPI spec version: 1.0.0
 */
import type { Order } from './order';

export interface OrderListResponse {
  data: Order[];
  total: number;
  page: number;
  pageSize: number;
}