# uozhou/packages/api-client/src/generated/model/orderWithItems.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Restaurant Operations API
 * Odyssey 餐厅运营系统 API - JWT 认证
 * OpenAPI spec version: 1.0.0
 */
import type { Customer } from './customer';
import type { Order } from './order';
import type { OrderItem } from './orderItem';

export type OrderWithItems = Order & ({
  items?: OrderItem[];
  customer?: Customer | null;
});