# uozhou/packages/api-client/src/generated/model/user.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Restaurant Operations API
 * Odyssey 餐厅运营系统 API - JWT 认证
 * OpenAPI spec version: 1.0.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  phone: string;
  name: string;
  /** @nullable */
  avatar?: string | null;
  role: UserRole;
  /** @nullable */
  customerId?: number | null;
}